// src/app/shared/models/Flight.ts
export interface Flight {
    aircraft: string;
    airline: string;
    arrivalTime: string;
    date: string;
    departureCity: string;
    departureTime: string;
    destinationCity: string;
    flightDuration: string;
    flightId: string;
    price: number;
    vacantFirstClassSeats: number;
    vacantSecondClassSeats: number;
}